pushing up a waiting page, working routing, and refactored structure
