+++ src/InsightLines.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import * as d3 from 'd3';
 
-import './styles/InsightLines.css';
+import './static/styles/InsightLines.css';
 
 class InsightLines extends Component {
 	constructor(props) {
