+++ src/LandingPage.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class LandingPage extends Component {
+  render() {
+    return (
+      <div className="landing-page">
+        Hello
+      </div>
+    );
+  }
+}
+
+export default LandingPage;
