+++ src/index.js
@@ -1,8 +1,15 @@
 import React from 'react';
+import { BrowserRouter } from 'react-router-dom';
 import ReactDOM from 'react-dom';
-import './styles/index.css';
+import './static/styles/index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(
+	(
+	<BrowserRouter>
+		<App />
+	</BrowserRouter>
+), document.getElementById('root')
+);
 registerServiceWorker();
