+++ src/static/styles/Album.css
@@ -7,8 +7,8 @@
   background-size: cover;
   background-position: center; }
   .album .album-text-wrapper {
-    background-color: rgba(255, 255, 255, 0.7);
-    color: black;
+    background-color: rgba(0, 0, 0, 0.4);
+    color: #ffffff;
     padding: 8px;
     width: 384px;
     position: absolute;
