+++ src/static/styles/Navigation.css
@@ -1,0 +1,22 @@
+.nav {
+  height: 64px;
+  background: #5392ff; }
+  .nav .nav-button {
+    display: inline-block;
+    height: 100%;
+    font-size: 24px;
+    text-decoration: none;
+    color: #ffffff;
+    border: none;
+    background: none;
+    border-right: 1px solid #ffffff; }
+    .nav .nav-button:not(:first-child) {
+      border-left: 1px solid #ffffff; }
+    .nav .nav-button:hover, .nav .nav-button:focus {
+      background: #2d74da;
+      cursor: pointer; }
+    .nav .nav-button .nav-button-content {
+      width: 64px;
+      height: 34px;
+      padding: 12px;
+      margin-top: 6px; }
