+++ src/static/styles/Timeline.css
@@ -1,3 +1,10 @@
+.timeline-root {
+  padding: 64px;
+  margin: 0; }
+  @media (max-width: 650px) {
+    .timeline-root {
+      padding: 16px; } }
+
 .timeline-base .row {
   width: 100%;
   position: relative;
@@ -9,7 +16,7 @@
   .timeline-base .row.line {
     height: 8px;
     border-radius: 8px;
-    background: #ffffff;
+    background: #000000;
     width: 100%;
     margin: 0; }
   .timeline-base .row.years {
@@ -17,7 +24,6 @@
     .timeline-base .row.years .year {
       font-size: 16px;
       height: 24px;
-      color: #ffffff;
       position: absolute;
       font-weight: 600; }
       .timeline-base .row.years .year .year-text {
@@ -29,7 +35,7 @@
         position: relative;
         border-radius: 4px;
         top: -2px;
-        background: white; }
+        background: #000000; }
       @media (max-width: 650px) {
         .timeline-base .row.years .year {
           font-size: 12px;
@@ -49,7 +55,6 @@
       margin-top: 35px; }
   .timeline-base .column.line {
     width: 8px;
-    background-color: #ffffff;
     border-radius: 16px; }
   .timeline-base .column.years {
     width: calc(30% - 4px); }
