+++ src/static/styles/TimelineWIP.css
@@ -1,0 +1,53 @@
+.wip-page {
+  height: calc(100vh - 64px);
+  background: #000000;
+  display: flex;
+  color: #ffffff; }
+  .wip-page .content {
+    margin: auto;
+    width: 432px; }
+    .wip-page .content > * {
+      margin-left: auto;
+      margin-right: auto;
+      display: block; }
+    .wip-page .content .gif {
+      width: 256px; }
+    .wip-page .content .loading-description {
+      position: relative;
+      top: -32px;
+      text-align: center; }
+      .wip-page .content .loading-description .sorry {
+        margin: 0;
+        margin-bottom: 12px;
+        font-size: 150px; }
+      .wip-page .content .loading-description .desc {
+        margin: 0;
+        font-size: 24px; }
+    .wip-page .content .submit-email {
+      width: 245px;
+      position: relative;
+      top: -18px; }
+      .wip-page .content .submit-email > * {
+        display: inline-block; }
+      .wip-page .content .submit-email .input-email {
+        height: 32px;
+        width: 196px;
+        border: none;
+        font-size: 18px;
+        padding-left: 14px;
+        line-height: 18; }
+      .wip-page .content .submit-email .submit-button {
+        display: inline-block;
+        height: 34px;
+        line-height: 34px;
+        width: 34px;
+        text-align: center;
+        vertical-align: bottom;
+        background-color: #5292ff; }
+    @media (max-width: 650px) {
+      .wip-page .content {
+        width: 300px; }
+        .wip-page .content .loading-description .sorry {
+          font-size: 72px; }
+        .wip-page .content .loading-description .desc {
+          font-size: 18px; } }
