+++ src/static/styles/sass/InsightLines.scss
@@ -15,7 +15,7 @@
 		bottom: 10px;
 
 		text {
-			fill: white;
+			// fill: $white;
 		}
 
 		.clear {
@@ -35,7 +35,7 @@
 		}
 
 		.circle {
-			fill: $white;
+			// fill: $white;
 		}
 	}
 }
