+++ src/static/styles/sass/TimelineBase.scss
@@ -1,0 +1,106 @@
+@import './utils.scss';
+
+$line-width: 8px;
+$percentiles-width: 30%;
+$album-width: 100% - $percentiles-width;
+
+// NOTE: $mobile-width: 650px;
+
+.timeline-root {
+	padding: $padding;
+	margin: 0;
+
+	// background: $black;
+	// color: $white;
+
+	@media (max-width: $mobile-width) {
+		padding: $padding-mobile;
+	}
+}
+
+.timeline-base {
+
+	.row {
+		width: 100%;
+		position: relative;
+		width: calc(100% - 64px);
+		margin: 0 32px;
+
+		&.albums {
+			overflow-x: hidden;
+			height: 400px;
+		}
+
+		&.line {
+			height: $line-width;
+			border-radius: $line-width;
+			// background: $white;
+			background: $black;
+			width: 100%;
+			margin: 0;
+		}
+
+		&.years {
+			height: 24px;
+
+			.year {
+				font-size: 16px;
+				height: 24px;
+				// color: $white;
+				position: absolute;
+				font-weight: 600;
+
+				.year-text {
+					position: relative;
+					right: 16px;
+				}
+
+				.year-tick {
+					height: 16px;
+					width: 4px;
+					position: relative;
+					border-radius: 4px;
+					top: -2px;
+					// background: white;
+					background: $black
+				}
+
+				@media (max-width: $mobile-width) {
+					font-size: 12px;
+					margin-right: 4px;
+				}
+			}
+		}
+
+		&.insights {
+			height: 400px;
+		}
+	}
+
+	.column {
+		display: inline-block;
+		height: 100%;
+		vertical-align: top;
+		position: relative;
+
+		&.percentiles {
+			width: calc(#{$album-width} - #{$line-width / 2});
+			overflow-y: hidden;
+
+			.albums {
+				margin-top: 35px;
+			}
+		}
+
+		&.line {
+			width: $line-width;
+			// background-color: $white;
+			border-radius: 16px;
+		}
+
+		&.years {
+			width: calc(#{$percentiles-width} - #{$line-width / 2});
+
+		}
+	}
+}
