+++ src/static/styles/sass/TimelineHeader.scss
@@ -13,7 +13,7 @@
 	}
 
 	.title {
-		color: $white;
+		// color: $white;
 		font-size: 72px;
 		margin: 0;
 	}
@@ -21,6 +21,7 @@
 	.header-image {
 		float: right;
 		width: 400px;
+		border-radius: 50%;
 
 		@media (max-width: $mobile-width) {
 			width: 100%;
