+++ src/static/styles/sass/TimelineWIP.scss
@@ -1,0 +1,84 @@
+@import './utils.scss';
+
+.wip-page {
+	height: calc(100vh - #{$nav-height});
+	background: $black;
+	display: flex;
+	color: $white;
+
+	.content {
+		margin: auto;
+		width: 432px;
+
+		> * {
+			margin-left: auto;
+			margin-right: auto;
+			display: block;
+		}
+
+		.gif {
+			width: 256px;
+
+		}
+
+		.loading-description {
+			position: relative;
+			top: -32px;
+			text-align: center;
+
+			.sorry {
+				margin: 0;
+				margin-bottom: 12px;
+				font-size: 150px;
+			}
+
+			.desc {
+				margin: 0;
+				font-size: 24px;
+			}
+		}
+
+		.submit-email {
+			width: 245px;
+			position: relative;
+			top: -18px;
+
+			> * {
+				display: inline-block;
+			}
+
+			.input-email {
+				height: 32px;
+				width: 196px;
+
+				border: none;
+				font-size: 18px;
+				padding-left: 14px;
+				line-height: 18;
+			}
+
+			.submit-button {
+				display: inline-block;
+				height: 34px;
+				line-height: 34px;
+				width: 34px;
+				text-align: center;
+				vertical-align: bottom;
+				background-color: #5292ff;
+			}
+		}
+
+		@media (max-width: $mobile-width) {
+			width: 300px;
+
+			.loading-description {
+				.sorry {
+					font-size: 72px;
+				}
+				.desc {
+					font-size: 18px;
+				}
+			}
+		}
+	}
+}
