+++ src/static/styles/sass/index.scss
@@ -1,0 +1,7 @@
+@import './utils.scss';
+
+body {
+	margin: 0;
+	padding: 0;
+	font-family: sans-serif;
+}
