+++ src/static/styles/sass/utils.scss
@@ -7,3 +7,5 @@ $padding-mobile: 16px;
 $scrollbar-width: 8px;
 
 $title-width: 200px;
+
+$nav-height: 64px;
